% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-functions.R
\name{bintest}
\alias{bintest}
\title{Tests for binary outcomes}
\usage{
bintest(binary, treat, var_est = "Unpooled")
}
\arguments{
\item{binary}{0=nonresponse, 1=response.}

\item{treat}{The treatment-group indicator, normally 0=control, 1=intervention.}

\item{var_est}{variance estimator: Unpooled/Pooled}
}
\value{
List: standardized statistic, statistic and variance.
}
\description{
performs a test for testing the null hypothesis that the proportions (probabilities of success) in two groups are the same.
}
\author{
Marta Bofill Roig
}
