% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-functions.R
\name{cop2}
\alias{cop2}
\title{Simulation binary and time-to-event data}
\usage{
cop2(v1, v2, theta)
}
\arguments{
\item{a.shape}{shape parameter Weibull}

\item{b.scale}{scale parameter Weibull}

\item{rate.param}{Censoring distribution parameter (Rate for the exponential, Max for Uniform(0,max))}

\item{prob0}{probability binary outcome}

\item{ass.par}{Association between binary and time-to-event outcome according to a Frank Copula}

\item{ss}{Sample size per arm}

\item{censoring}{Censoring distribution. Options: "Exp": Exponential; "Unif": Uniform}
}
\value{
Binary and time-to-event data
}
\description{
simulates two-sample binary and time-to-event dataset.
}
\references{
Trivedi et al, Appendix A
}
\author{
Marta Bofill Roig
}
