% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstats_boots-functions.R
\name{lstats_boots}
\alias{lstats_boots}
\title{L-Statistics for binary and time-to-event outcomes (bootstrap variance)}
\usage{
lstats_boots(
  time,
  status,
  binary,
  treat,
  tau,
  rho = 0,
  gam = 0,
  eta = 1,
  wb = 0.5,
  ws = 0.5,
  Boot = 50
)
}
\arguments{
\item{time}{The observed time.}

\item{status}{The status indicator, normally 0=alive, 1=dead.}

\item{binary}{0=nonresponse, 1=response.}

\item{treat}{The treatment-group indicator, normally 0=control, 1=intervention.}

\item{tau}{follow-up. Default NULL denoting the last time in which both groups had patients at risk.}

\item{rho}{A scalar parameter that controls the type of test (see Weights).}

\item{gam}{A scalar parameter that controls the type of test (see Weights).}

\item{eta}{A scalar parameter that controls the type of test (see Weights).}

\item{wb}{A scalar parameter that controls the type of test (see Weights).}

\item{ws}{A scalar parameter that controls the type of test (see Weights).}

\item{Boot}{number of bootstrap samples (default Boot=50).}
}
\value{
List: standardized statistic, statistic and variance.
}
\description{
performs a L-statistic for right-censored and binary data (bootstrap variance estimator).
}
\author{
Marta Bofill Roig
}
