% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-functions.R
\name{survtest}
\alias{survtest}
\title{Tests for time-to-event outcomes}
\usage{
survtest(
  time,
  status,
  treat,
  tau = NULL,
  rho = 0,
  gam = 0,
  eta = 1,
  var_est = "Unpooled"
)
}
\arguments{
\item{time}{The observed time.}

\item{status}{The status indicator, normally 0=alive, 1=dead.}

\item{treat}{The treatment-group indicator, normally 0=control, 1=intervention.}

\item{tau}{follow-up. Default NULL denoting the last time in which both groups had patients at risk.}

\item{rho}{A scalar parameter that controls the type of test (see Weights).}

\item{gam}{A scalar parameter that controls the type of test (see Weights).}

\item{eta}{A scalar parameter that controls the type of test (see Weights).}

\item{var_est}{indicates the variance estimate to use ('Pooled' or 'Unpooled')}
}
\value{
List: standardized statistic, statistic and variance.
}
\description{
performs a test for right-censored data. It uses the Weighted Kaplan-Meier family of statistics for testing the differences of two survival curves.
}
\author{
Marta Bofill Roig
}
